refactor(admin): extract Settings > Checkout settings into standalone component
Behaviour-preserving Tier-0 decomposition of the 19k-line admin.component.ts.
Move the checkout-settings panel (template + checkoutSettingsForm state +
loadCheckoutSettings/saveCheckoutSettings) into a new standalone OnPush
component (app-admin-checkout-settings) under pages/admin/settings/. The shared
CMS content-version bookkeeping stays on AdminComponent and is threaded into the
child via the four stable bound callbacks (rememberContentVersion,
withExpectedVersion, handleContentConflict, forgetContentVersion) so all CMS
panels keep sharing one contentVersions map. The parent template now renders
the host with *ngIf="section()==='settings'"; the child self-loads on init. The
load-error path now calls the injected forgetContentVersion('site.checkout')
(same effect as the prior direct contentVersions delete, matching the
company-info pattern).

Tests: checkout behaviour/branch coverage moved out of admin.component.spec and
admin.component.coverage.spec into the new admin-checkout-settings.component.spec;
loadCheckoutSettings dropped from the settings-load spy lists. Build clean; 620
targeted specs green (4 netflow + admin.component + coverage + new spec); settings
section screenshot confirms the checkout panel renders unchanged with no layout
breakage.
